+++ MicrosoftDefender/microsoftdefender_modules/client/auth.py
@@ -30,18 +30,10 @@ class MicrosoftDefenderCredentials:
 
 
 class ApiKeyAuthentication(AuthBase):
-    def __init__(
-        self,
-        app_id: str,
-        app_secret: str,
-        tenant_id: str,
-        ratelimit_per_minute: int,
-        base_url: str = "https://api.securitycenter.microsoft.com",
-    ):
+    def __init__(self, app_id: str, app_secret: str, tenant_id: str, ratelimit_per_minute: int):
         self.__app_id = app_id
         self.__app_secret = app_secret
         self.__tenant_id = tenant_id
-        self.__base_url = base_url.rstrip("/")
 
         self.__api_credentials: MicrosoftDefenderCredentials | None = None
 
@@ -62,8 +54,9 @@ class ApiKeyAuthentication(AuthBase):
 
         if self.__api_credentials is None or current_dt + timedelta(seconds=300) >= self.__api_credentials.expires_at:
             url = "https://login.microsoftonline.com/%s/oauth2/token" % self.__tenant_id
+            resource_app_id_uri = "https://api.securitycenter.microsoft.com"
             body = {
-                "resource": self.__base_url,
+                "resource": resource_app_id_uri,
                 "client_id": self.__app_id,
                 "client_secret": self.__app_secret,
                 "grant_type": "client_credentials",